core-embed: add variation fallback. refact.
