+++ extensions/extended-blocks/core-embed/index.js
@@ -1,31 +1,5 @@
+
 /**
- * WordPress dependencies
+ * Internal dependencies
  */
-import { registerBlockVariation } from '@wordpress/blocks';
-import { Path, SVG } from '@wordpress/components';
-import { getIconColor } from '../../shared/block-icons';
-import { __ } from '@wordpress/i18n';
-
-const embedLoomIcon = {
-	foreground: getIconColor(),
-	src: (
-		<SVG viewBox="0 0 48 48">
-			<Path d="M48 24.1194C48 17.52 44.22 11.97 38.6866 9.31343C36.03 3.63 30.3 0 24.1194 0C17.52 0 11.97 3.78 9.31343 9.31343C3.63 11.97 0 17.7 0 24.1194C0 30.48 3.78 36.03 9.31343 38.6866C11.97 44.37 17.7 48 24.1194 48C30.48 48 36.03 44.22 38.6866 38.6866C44.37 36.03 48 30.3 48 24.1194ZM28.8955 18.8657C29.4925 20.4776 29.8507 22.1493 29.8507 23.8806C29.8507 25.6418 29.4925 27.3134 28.8955 28.8955C27.2836 29.4925 25.6119 29.8507 23.8806 29.8507C22.1194 29.8507 20.4478 29.4925 18.8657 28.8955C18.2687 27.2836 17.9104 25.6119 17.9104 23.8806C17.9104 22.1194 18.2687 20.4478 18.8657 18.8657C20.4776 18.2687 22.1493 17.9104 23.8806 17.9104C25.6418 17.9104 27.2836 18.2687 28.8955 18.8657ZM24.1194 3.10448C27.9735 3.10448 31.6671 4.84788 34.1493 7.8806C33.2628 7.62551 32.3468 7.53686 31.5224 7.64179C24.8412 7.53686 19.1973 11.4669 16.7164 17.194C16.597 17.1404 16.597 17.1404 16.7164 17.194C15.681 19.2088 15.1787 21.425 15.2836 23.8806C15.1787 24.705 15.2673 25.621 15.5224 26.5075C12.5192 24.0253 10.7463 20.3317 10.7463 16.4776C10.7463 9.07342 16.7152 3.10448 24.1194 3.10448ZM3.10448 23.8806C3.10448 20.0265 4.84788 16.3329 7.8806 13.8507C7.62551 14.7372 7.53686 15.6532 7.64179 16.4776C7.53686 23.1588 11.4669 28.8027 17.194 31.2836C17.1404 31.403 17.1404 31.403 17.194 31.2836C19.2088 32.319 21.425 32.8213 23.8806 32.7164C24.705 32.8213 25.621 32.7327 26.5075 32.4776C24.0253 35.4808 20.3317 37.2537 16.4776 37.2537C9.07342 37.2537 3.10448 31.2848 3.10448 23.8806ZM23.8806 44.4179C20.0265 44.4179 16.3329 42.6745 13.8507 39.6418C14.7372 39.8969 15.6532 39.9855 16.4776 39.8806C23.1588 39.9855 28.8027 36.0555 31.2836 30.3284C31.403 30.382 31.403 30.382 31.2836 30.3284C32.319 28.3136 32.8213 26.0974 32.7164 23.6418C32.8213 22.8174 32.7327 21.9014 32.4776 21.0149C35.4808 23.4971 37.2537 27.1907 37.2537 31.0448C37.2537 38.449 31.2848 44.4179 23.8806 44.4179ZM30.3284 16.7164C30.382 16.597 30.382 16.597 30.3284 16.7164C28.3136 15.681 26.0974 15.1787 23.6418 15.2836C22.8174 15.1787 21.9014 15.2673 21.0149 15.5224C23.4971 12.5192 27.1907 10.7463 31.0448 10.7463C38.449 10.7463 44.4179 16.7152 44.4179 24.1194C44.4179 27.9735 42.6745 31.6671 39.6418 34.1493C39.8969 33.2628 39.9855 32.3468 39.8806 31.5224C39.9855 24.8412 36.0555 19.1973 30.3284 16.7164Z" />
-		</SVG>
-	),
-};
-
-// The `core/embed` block started using block variations in WP 5.5.
-// Previous WP versions will not see the Loom block variation in the inserter, but can still use the generic embed block,
-// with Loom registered as an oembed provider.
-if ( typeof registerBlockVariation !== undefined ) {
-	registerBlockVariation( 'core/embed', {
-		name: 'loom',
-		title: 'Loom',
-		icon: embedLoomIcon,
-		keywords: [ __( 'video' ) ],
-		description: __( 'Embed a Loom video.' ),
-		patterns: [ /^https?:\/\/(www\.)?loom\.com\/share\/.+/i ],
-		attributes: { providerNameSlug: 'loom', responsive: true },
-	} );
-}
+import './variations';
